feat: Add `course_revision` number as `Course Version` to the outline page.
Needed a way to apply a version number to a course to ensure that the learner has a specific version of the course.

This was a request from the `Choose Aerospace` administers to ensure that the schools are using the latest version.

Adding this number into the course is found on the Schedule & Details page Course Details section (new field highlighted below).