feat: Indent `subsection` sequences on the course outline.
Verified that the instructional design team would like the the `subsection` sequences indented. This will allow for visual separation of the content and provide better nesting.

I tried using bootstrap spacing https://getbootstrap.com/docs/4.3/utilities/spacing/ of `pl-5` but it was too far in for the indent and I needed to specify an exact value to align the checkmark progress with the title of the section name.

Ref: https://github.com/CUCWD/frontend-app-learning/issues/25