feat(badges): Initial step setting up the BadgeProgressTab to adhere …
…to the new frontend structure.

Todo: This just renders the start of BadgeProgressTab at this URL, however, we still need to render out the progress details.

Also the Jest tests for `Test fetchBadgesProgressTab` still need to be worked out. The `Should fetch, normalize, and save metadata` is not working due to the `frontend-app-learning/src/course-home/data/__factories__/badgeProgressTabData.factory.js` Factory needing to be setup.

http://localhost:2000/course/{courseId}/badges/progress